First working solution.

+++ hourly_archivizer.ps1
@@ -1,7 +1,7 @@
 # Get environment from config file
-$configfile = (Split-Path $MyInvocation.MyCommand.Path -Parent) + '\archivizer.conf'
+$currentpath = (Split-Path $MyInvocation.MyCommand.Path -Parent)
 
-Foreach ($i in $(Get-Content $configfile -Encoding "UTF8")){
+Foreach ($i in $(Get-Content "$currentpath\archivizer.conf" -Encoding "UTF8")){
     Set-Variable -Name $i.split("=")[0] -Value $i.split("=",2)[1]
 }
 
@@ -26,20 +26,27 @@ Function Merge {
     [Parameter(Mandatory=$true)][string]$Source,
     [Parameter(Mandatory=$true)][string]$Destination
   )
-  Get-ChildItem -Path $Source | ForEach-Object {
-    if (Test-Path -Path "$Destination\$_") {
-      if ((get-item "$Source\$_").PSIsContainer) {
-        Merge -Source "$Source\$_" -Destination "$Destination\$_"
-        Remove-Item "$Source\$_"
+  if (Test-Path -Path "$Destination") {
+    Get-ChildItem -Path $Source | ForEach-Object {
+      if (Test-Path -Path "$Destination\$_") {
+        if ((get-item "$Source\$_").PSIsContainer) {
+          Merge -Source "$Source\$_" -Destination "$Destination\$_"
+          Remove-Item "$Source\$_"
+          }
+        else {
+          Move-Item -LiteralPath $Source\$_ -Destination $Destination -Force
+          }
         }
       else {
         Move-Item -LiteralPath $Source\$_ -Destination $Destination -Force
-        }
+        Writelog("$Source\$_ has been moved to $Destination.")
+        }    
       }
+    }
     else {
-      Move-Item -LiteralPath $Source\$_ -Destination $Destination -Force
-      }    
-    }  
+      Move-Item -LiteralPath $Source -Destination $Destination
+      Writelog("$Source has been moved to $Destination.")        
+    }
 }
 
 Function CreateIfNotExist {
@@ -73,7 +80,7 @@ Function RarAndLog {
     WriteLog("$Source does not exist so cannot be archivized")
     }
   else {
-    $argList = @("a",  "-r", "-ep1", "$Destination.rar" ,"$Source\*.*")
+    $argList = @("a",  "-r", "-ep1", "-ri14", "$Destination.rar" ,"$Source\*.*")
     Start-Process -FilePath "C:\Program Files\Winrar\winrar.exe" -ArgumentList $argList -NoNewWindow -Wait
     WriteLog("Content of $Source has been archivzed to $Destination.rar file.")
   }  
@@ -88,7 +95,14 @@ Function MoveRarMove {
 MoveAndLog -Source "$FromFolderPath\$Day" -Destination $Destination
 RarAndLog -Source $Destination -Destination $Destination\$PcName@$Day_HH
 MoveAndLog -Source "$Destination\$PcName@$Day_HH.rar" -Destination "$ToFolderPath\$Day"
+}
 
+
+Function MergeBackAndLog {
+  Get-ChildItem -Path "$FromFolderPath\Temporary\$Yesterday" | ForEach-Object {
+    Merge -Source "$FromFolderPath\Temporary\$Yesterday\$_\$Yesterday" -Destination "$FromFolderPath\$Yesterday"
+    WriteLog("$FromFolderPath\Temporary\$Yesterday\$_\$Yesterday has been successfully merged into $FromFolderPath.")
+  }
 }
 
 WriteLog("******************************************************************************************")
@@ -96,14 +110,15 @@ WriteLog("Script has started...")
 
 $Today = (Get-Date).toString('yyyy_MM_dd')
 $Today_HH = (Get-Date).toString('yyyy_MM_dd_HH')
-$Destination = "$FromFolderPath\Temporary\$Today_HH"
+$Destination = "$FromFolderPath\Temporary\$Today\$Today_HH"
 
 MoveRarMove -Day $Today -Day_HH $Today_HH -Destination $Destination
 
 if ($(Get-Date).Hour -eq 0){  
   $Yesterday = (Get-Date).AddDays(-1).toString('yyyy_MM_dd')
   $Yesterday_HH = $Yesterday+"_24"
-  MoveRarMove -Day $Yesterday -Day_HH $Yesterday_HH -Destination "$FromFolderPath\Temporary\$Yesterday_HH"
+  MoveRarMove -Day $Yesterday -Day_HH $Yesterday_HH -Destination "$FromFolderPath\Temporary\$Yesterday\$Yesterday_HH"
+  MergeBackAndLog
 }
 
 WriteLog("Script completed.")
